renamed src/client/store/shared to src/client/store/APIRequestsHandling
