+++ src/client/Components/LoggedInUserController.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { Route, Switch, Redirect } from 'react-router'
 import Home from './Home/'
 import LanguagePage from './LanguagePage'
-import { userLoggedInThunk } from '../store/shared/thunks'
+import { userLoggedInThunk } from '../store/APIRequestsHandling/thunks'
 import { User } from '../../entities'
 import { ReduxState } from '../store'
 import { IUserReducerDataSlice } from '../store/user/reducer'
@@ -11,7 +11,7 @@ import { IUserLoggedInLanguagesDataSlice } from '../store/userlanguage/reducer'
 import { IChatGroupReducerState } from '../store/chatgroup/reducer'
 import { IUserChatGroupReducerState } from '../store/userchatgroup/reducer'
 import { IAuthReducerState } from '../store/auth/reducer'
-import { OnApiFailureActionTypes } from '../store/shared/types'
+import { OnApiFailureActionTypes } from '../store/APIRequestsHandling/types'
 import CircularProgress from '@material-ui/core/CircularProgress'
 //import Backdrop from '@material-ui/core/Backdrop'
 import Typography from '@material-ui/core/Typography'
