+++ src/client/Components/Login_Signup/Login.tsx
@@ -9,7 +9,7 @@ import useStyles from './styles'
 import { connect } from 'react-redux'
 import { loginUserProcess } from '../../store/auth/thunks'
 import { IUserSignInDTO } from '../../../server/auth/auth.dto'
-import { OnApiFailureActionTypes } from '../../store/shared/types'
+import { OnApiFailureActionTypes } from '../../store/APIRequestsHandling/types'
 import { ReduxState } from '../../store'
 import { IAuthReducerState } from '../../store/auth/reducer'
 import Grid from '@material-ui/core/Grid'
