+++ src/client/store/APIRequestsHandling/helperfunctions.ts
