+++ src/client/store/language/actions.ts
@@ -5,7 +5,7 @@ import {
   RequestDataConstants,
   OnApiFailureActionTypes,
   ENTERED_SITE_BASE_DATA_RECEIVED
-} from '../shared/types'
+} from '../APIRequestsHandling/types'
 const { ENTERED_SITE_LOADING_BASE_DATA_REQUEST } = RequestDataConstants
 const {
   ENTERED_SITE_LOADING_BASE_DATA_REQUEST_FAILED
